requirements.txt
aiogram==3.22.0
python-dotenv==1.1.0
pandas==2.3.3
matplotlib==3.10.8
yfinance==0.2.66
statsmodels==0.14.5
scikit-learn==1.7.2
torch~=2.9.0
scipy~=1.16.3